srb 10.9

+++ service-core/src/main/java/com/atguigu/srb/core/service/BorrowerAttachService.java
@@ -16,5 +16,5 @@ import java.util.List;
  */
 public interface BorrowerAttachService extends IService<BorrowerAttach> {
 
-    List<BorrowerAttachVO> selectBorrowerAttachVOList(Integer id);
+    List<BorrowerAttachVO> selectBorrowerAttachVOList(Long id);
 }
